pyproject.toml
[project]
name = "django-base"
version = "1.0.0"
description = "A django project for learning propouses"
authors = [
    { name = "Matheus Santos", email = "maasantos776@gmail.com" }
]
dependencies = [
    "Django",
    "python-dotenv",
    "pillow",
    "Faker",
    "django-select2",
    "djangorestframework",
    "djangorestframework-simplejwt",
    "setuptools",
    "wheel",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-django",
    "coverage",
    "parameterized",
    "selenium"
]
dev = [
    "django-debug-toolbar",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "static",
    "templates",
    "static_base"
#    "core",
#    "authors",
#    "recipes",
]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["static/**", "templates/**"]
